tests/wpt/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.zerosource.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.zerosource</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.zerosource</h1>
<p class="desc">drawImage with zero-sized source rectangle draws nothing without exception</p>


<script>
var t = async_test("drawImage with zero-sized source rectangle draws nothing without exception");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
var promise = new Promise(function(resolve, reject) {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", '/images/green.png');
    xhr.responseType = 'blob';
    xhr.send();
    xhr.onload = function() {
        resolve(xhr.response);
    };
});
promise.then(function(response) {
    createImageBitmap(response).then(bitmap => {
        ctx.drawImage(bitmap, 10, 10, 0, 1, 0, 0, 100, 50);
        ctx.drawImage(bitmap, 10, 10, 1, 0, 0, 0, 100, 50);
        ctx.drawImage(bitmap, 10, 10, 0, 0, 0, 0, 100, 50);
        _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);
    }, t_fail);
}).then(t_pass, t_fail);

});
</script>
