LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.zerosource-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: IndexSizeError: The index is not in the allowed range.
2d.drawImage.zerosource

drawImage with zero-sized source rectangle draws nothing without exception


PASS drawImage with zero-sized source rectangle draws nothing without exception
